handle polygons all places coords in extent check
